refactor(#283): 시니어·보호자 홈 서비스 중복 로직 공통 컴포넌트로 분리
- FamilyMemberQueryService — 가족 구성원 ID 수집 로직 분리
- HomeAlbumRecommendationService — 앨범 추천 점수 계산 로직 분리
- GuardianHomeService·SeniorHomeService 중복 코드 제거 후 공통 컴포넌트 주입으로 교체
- 홈 서비스 단위 테스트 수정
