+++ backend/widyu-api/src/main/java/com/widyu/home/application/GuardianHomeService.java
@@ -1,7 +1,5 @@
 package com.widyu.home.application;
 
-import com.widyu.album.Album;
-import com.widyu.album.repository.AlbumRepository;
 import com.widyu.global.entity.Status;
 import com.widyu.global.error.BusinessException;
 import com.widyu.global.error.ErrorCode;
@@ -12,7 +10,6 @@ import com.widyu.goal.medicineschedule.repository.MedicationProofRepository;
 import com.widyu.goal.medicineschedule.repository.MedicineScheduleRepository;
 import com.widyu.goal.walk.repository.WalkRepository;
 import com.widyu.healthschedule.HealthSchedule;
-import com.widyu.heart.HeartRateResult;
 import com.widyu.heart.repository.HeartRateResultRepository;
 import com.widyu.home.dto.response.GuardianHomeCardsResponse;
 import com.widyu.home.dto.response.GuardianSeniorListResponse;
@@ -35,7 +32,6 @@ import java.util.Optional;
 import java.util.Set;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
-import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -43,12 +39,6 @@ import org.springframework.transaction.annotation.Transactional;
 @RequiredArgsConstructor
 @Transactional(readOnly = true)
 public class GuardianHomeService {
-
-    private static final int ALBUM_CANDIDATE_SIZE = 10;
-    private static final int ALBUM_RESULT_SIZE = 3;
-    private static final int LIKE_WEIGHT = 3;
-    private static final int COMMENT_WEIGHT = 2;
-    private static final int DATE_BONUS = 10;
 
     private final MemberUtil memberUtil;
     private final MemberRepository memberRepository;
@@ -59,7 +49,7 @@ public class GuardianHomeService {
     private final MedicationProofRepository medicationProofRepository;
     private final HealthScheduleRepository healthScheduleRepository;
     private final WalkRepository walkRepository;
-    private final AlbumRepository albumRepository;
+    private final HomeAlbumRecommendationService albumRecommendationService;
 
     public GuardianHomeCardsResponse getHomeCards(Long memberId) {
         Member guardian = memberUtil.getCurrentMember();
@@ -166,60 +156,13 @@ public class GuardianHomeService {
     }
 
     private List<GuardianHomeCardsResponse.AlbumInfo> getScoredAlbums(Member senior, LocalDate today) {
-        List<Long> familyMemberIds = getFamilyMemberIds(senior);
-
-        List<Long> candidateIds = albumRepository
-                .findTopScoredAlbumIdsByMemberIds(familyMemberIds, PageRequest.of(0, ALBUM_CANDIDATE_SIZE))
-                .getContent();
-
-        if (candidateIds.isEmpty()) {
-            return List.of();
-        }
-
-        return albumRepository.findAlbumsWithCollectionsByIds(candidateIds).stream()
-                .sorted(Comparator.comparingInt((Album album) -> calculateScore(album, today)).reversed())
-                .limit(ALBUM_RESULT_SIZE)
+        return albumRecommendationService.recommendAlbums(senior, today).stream()
                 .map(GuardianHomeCardsResponse.AlbumInfo::from)
                 .toList();
     }
 
-    private List<Long> getFamilyMemberIds(Member senior) {
-        SeniorProfile seniorProfile = seniorProfileRepository.findByMemberId(senior.getId())
-                .orElse(null);
-
-        if (seniorProfile == null) {
-            return List.of(senior.getId());
-        }
-
-        Long familyId = seniorProfile.getFamily().getId();
-
-        List<Long> seniorIds = seniorProfileRepository.findAllByFamilyId(familyId).stream()
-                .map(sp -> sp.getMember().getId())
-                .toList();
-
-        List<Long> guardianIds = familyMembershipRepository.findAllByFamilyIdWithGuardian(familyId).stream()
-                .map(fm -> fm.getGuardian().getId())
-                .toList();
-
-        return java.util.stream.Stream.concat(seniorIds.stream(), guardianIds.stream()).toList();
-    }
-
-    private int calculateScore(Album album, LocalDate today) {
-        int baseScore = album.getLikeCount() * LIKE_WEIGHT + album.getCommentCount() * COMMENT_WEIGHT;
-        if (isAnniversary(album, today)) {
-            return baseScore + DATE_BONUS;
-        }
-        return baseScore;
-    }
-
-    private boolean isAnniversary(Album album, LocalDate today) {
-        LocalDate albumDate = album.getCreatedAt().toLocalDate();
-        return albumDate.getMonth() == today.getMonth()
-                && albumDate.getDayOfMonth() == today.getDayOfMonth();
-    }
-
     private GuardianHomeCardsResponse.HealthScheduleInfo getHealthScheduleInfo(Member senior, LocalDate today) {
-        return healthScheduleRepository.findByMemberIdAndDate(senior.getId(), today)
+        return healthScheduleRepository.findByMemberIdAndDate(senior.getId(), today.atStartOfDay(), today.plusDays(1).atStartOfDay())
                 .stream()
                 .min(Comparator.comparing(HealthSchedule::getScheduledAt))
                 .map(GuardianHomeCardsResponse.HealthScheduleInfo::from)
