+++ backend/widyu-api/src/test/java/com/widyu/home/GuardianHomeServiceTest.java
@@ -3,11 +3,10 @@ package com.widyu.home;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.anyList;
 import static org.mockito.BDDMockito.given;
 import static org.mockito.Mockito.mock;
 
-import com.widyu.album.repository.AlbumRepository;
+import com.widyu.album.Album;
 import com.widyu.global.entity.Status;
 import com.widyu.global.error.BusinessException;
 import com.widyu.global.util.MemberUtil;
@@ -17,6 +16,7 @@ import com.widyu.goal.medicineschedule.repository.MedicineScheduleRepository;
 import com.widyu.goal.walk.repository.WalkRepository;
 import com.widyu.heart.repository.HeartRateResultRepository;
 import com.widyu.home.application.GuardianHomeService;
+import com.widyu.home.application.HomeAlbumRecommendationService;
 import com.widyu.home.dto.response.GuardianHomeCardsResponse;
 import com.widyu.medicine.MedicineSchedule;
 import com.widyu.member.Family;
@@ -36,7 +36,6 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import org.springframework.data.domain.SliceImpl;
 
 @ExtendWith(MockitoExtension.class)
 @DisplayName("GuardianHomeService 단위 테스트")
@@ -53,7 +52,7 @@ class GuardianHomeServiceTest {
     @Mock private MedicationProofRepository medicationProofRepository;
     @Mock private HealthScheduleRepository healthScheduleRepository;
     @Mock private WalkRepository walkRepository;
-    @Mock private AlbumRepository albumRepository;
+    @Mock private HomeAlbumRecommendationService albumRecommendationService;
 
     @Test
     @DisplayName("시니어가 보호자 홈을 호출하면 403 예외가 발생한다")
@@ -103,13 +102,9 @@ class GuardianHomeServiceTest {
         given(seniorProfileRepository.findAllByFamilyIdWithMember(any())).willReturn(List.of(seniorProfile));
         given(heartRateResultRepository.findByMemberId(any())).willReturn(Optional.empty());
         given(medicineScheduleRepository.findByMemberAndStatusWithDetails(any(), any())).willReturn(List.of());
-        given(healthScheduleRepository.findByMemberIdAndDate(any(), any())).willReturn(List.of());
+        given(healthScheduleRepository.findByMemberIdAndDate(any(), any(), any())).willReturn(List.of());
         given(walkRepository.findByMemberAndWalkDate(any(), any())).willReturn(Optional.empty());
-        given(seniorProfileRepository.findByMemberId(any())).willReturn(Optional.of(seniorProfile));
-        given(seniorProfileRepository.findAllByFamilyId(any())).willReturn(List.of(seniorProfile));
-        given(familyMembershipRepository.findAllByFamilyIdWithGuardian(any())).willReturn(List.of(membership));
-        given(albumRepository.findTopScoredAlbumIdsByMemberIds(anyList(), any()))
-                .willReturn(new SliceImpl<>(List.of()));
+        given(albumRecommendationService.recommendAlbums(any(), any())).willReturn(List.of());
 
         // when
         GuardianHomeCardsResponse response = guardianHomeService.getHomeCards(null);
@@ -134,13 +129,9 @@ class GuardianHomeServiceTest {
         given(seniorProfileRepository.findAllByFamilyIdWithMember(any())).willReturn(List.of(seniorProfile));
         given(heartRateResultRepository.findByMemberId(any())).willReturn(Optional.empty());
         given(medicineScheduleRepository.findByMemberAndStatusWithDetails(any(), any())).willReturn(List.of());
-        given(healthScheduleRepository.findByMemberIdAndDate(any(), any())).willReturn(List.of());
+        given(healthScheduleRepository.findByMemberIdAndDate(any(), any(), any())).willReturn(List.of());
         given(walkRepository.findByMemberAndWalkDate(any(), any())).willReturn(Optional.empty());
-        given(seniorProfileRepository.findByMemberId(any())).willReturn(Optional.of(seniorProfile));
-        given(seniorProfileRepository.findAllByFamilyId(any())).willReturn(List.of(seniorProfile));
-        given(familyMembershipRepository.findAllByFamilyIdWithGuardian(any())).willReturn(List.of(membership));
-        given(albumRepository.findTopScoredAlbumIdsByMemberIds(anyList(), any()))
-                .willReturn(new SliceImpl<>(List.of()));
+        given(albumRecommendationService.recommendAlbums(any(), any())).willReturn(List.of());
 
         // when
         GuardianHomeCardsResponse response = guardianHomeService.getHomeCards(null);
@@ -165,13 +156,9 @@ class GuardianHomeServiceTest {
         given(seniorProfileRepository.findAllByFamilyIdWithMember(any())).willReturn(List.of(seniorProfile));
         given(heartRateResultRepository.findByMemberId(any())).willReturn(Optional.empty());
         given(medicineScheduleRepository.findByMemberAndStatusWithDetails(any(), any())).willReturn(List.of());
-        given(healthScheduleRepository.findByMemberIdAndDate(any(), any())).willReturn(List.of());
+        given(healthScheduleRepository.findByMemberIdAndDate(any(), any(), any())).willReturn(List.of());
         given(walkRepository.findByMemberAndWalkDate(any(), any())).willReturn(Optional.empty());
-        given(seniorProfileRepository.findByMemberId(any())).willReturn(Optional.of(seniorProfile));
-        given(seniorProfileRepository.findAllByFamilyId(any())).willReturn(List.of(seniorProfile));
-        given(familyMembershipRepository.findAllByFamilyIdWithGuardian(any())).willReturn(List.of(membership));
-        given(albumRepository.findTopScoredAlbumIdsByMemberIds(anyList(), any()))
-                .willReturn(new SliceImpl<>(List.of()));
+        given(albumRecommendationService.recommendAlbums(any(), any())).willReturn(List.of());
 
         // when
         GuardianHomeCardsResponse response = guardianHomeService.getHomeCards(null);
@@ -206,13 +193,9 @@ class GuardianHomeServiceTest {
                 .willReturn(List.of(schedule1, schedule2));
         given(medicationProofRepository.findByMemberIdAndDateRange(any(), any(), any()))
                 .willReturn(List.of());
-        given(healthScheduleRepository.findByMemberIdAndDate(any(), any())).willReturn(List.of());
+        given(healthScheduleRepository.findByMemberIdAndDate(any(), any(), any())).willReturn(List.of());
         given(walkRepository.findByMemberAndWalkDate(any(), any())).willReturn(Optional.empty());
-        given(seniorProfileRepository.findByMemberId(any())).willReturn(Optional.of(seniorProfile));
-        given(seniorProfileRepository.findAllByFamilyId(any())).willReturn(List.of(seniorProfile));
-        given(familyMembershipRepository.findAllByFamilyIdWithGuardian(any())).willReturn(List.of(membership));
-        given(albumRepository.findTopScoredAlbumIdsByMemberIds(anyList(), any()))
-                .willReturn(new SliceImpl<>(List.of()));
+        given(albumRecommendationService.recommendAlbums(any(), any())).willReturn(List.of());
 
         // when
         GuardianHomeCardsResponse response = guardianHomeService.getHomeCards(null);
