+++ backend/widyu-api/src/test/java/com/widyu/home/SeniorHomeServiceTest.java
@@ -3,34 +3,23 @@ package com.widyu.home;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.anyList;
 import static org.mockito.BDDMockito.given;
 import static org.mockito.Mockito.mock;
 
-import com.widyu.album.repository.AlbumRepository;
-import com.widyu.global.entity.Status;
 import com.widyu.global.error.BusinessException;
 import com.widyu.global.util.MemberUtil;
 import com.widyu.goal.healthschedule.repository.HealthScheduleRepository;
 import com.widyu.goal.medicineschedule.repository.MedicationProofRepository;
 import com.widyu.goal.medicineschedule.repository.MedicineScheduleRepository;
 import com.widyu.goal.walk.repository.WalkRepository;
-import com.widyu.healthschedule.HealthSchedule;
 import com.widyu.heart.repository.HeartRateResultRepository;
+import com.widyu.home.application.HomeAlbumRecommendationService;
 import com.widyu.home.application.SeniorHomeService;
 import com.widyu.home.dto.response.SeniorHomeCardsResponse;
 import com.widyu.medicine.MedicationProof;
 import com.widyu.medicine.MedicineSchedule;
-import com.widyu.member.Family;
-import com.widyu.member.FamilyMembership;
 import com.widyu.member.Member;
 import com.widyu.member.MemberType;
-import com.widyu.member.SeniorProfile;
-import com.widyu.member.repository.FamilyMembershipRepository;
-import com.widyu.member.repository.SeniorProfileRepository;
-import com.widyu.walk.Walk;
-import java.time.LocalDate;
-import java.time.LocalDateTime;
 import java.time.LocalTime;
 import java.util.List;
 import java.util.Optional;
@@ -40,7 +29,6 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import org.springframework.data.domain.SliceImpl;
 
 @ExtendWith(MockitoExtension.class)
 @DisplayName("SeniorHomeService 단위 테스트")
@@ -53,10 +41,8 @@ class SeniorHomeServiceTest {
     @Mock private MedicationProofRepository medicationProofRepository;
     @Mock private WalkRepository walkRepository;
     @Mock private HealthScheduleRepository healthScheduleRepository;
-    @Mock private AlbumRepository albumRepository;
     @Mock private HeartRateResultRepository heartRateResultRepository;
-    @Mock private SeniorProfileRepository seniorProfileRepository;
-    @Mock private FamilyMembershipRepository familyMembershipRepository;
+    @Mock private HomeAlbumRecommendationService albumRecommendationService;
 
     @Test
     @DisplayName("보호자가 시니어 홈을 호출하면 403 예외가 발생한다")
@@ -75,22 +61,13 @@ class SeniorHomeServiceTest {
     void 약_스케줄이_없으면_medicine은_null을_반환한다() {
         // given
         Member senior = Member.createMember(MemberType.SENIOR, "부모님", "01011112222");
-        Family family = Family.createFamily("FAMA01");
-        SeniorProfile seniorProfile = SeniorProfile.createSeniorProfile(
-                senior, family, "서울시", null, "1110001", null);
-        FamilyMembership membership = FamilyMembership.createLeaderMembership(family,
-                Member.createMember(MemberType.GUARDIAN, "보호자", "01099999999"));
 
         given(memberUtil.getCurrentMember()).willReturn(senior);
         given(medicineScheduleRepository.findByMemberAndStatusWithDetails(any(), any())).willReturn(List.of());
         given(heartRateResultRepository.findByMemberId(any())).willReturn(Optional.empty());
         given(healthScheduleRepository.findByMemberIdAndWeek(any(), any(), any())).willReturn(List.of());
         given(walkRepository.findByMemberAndWalkDate(any(), any())).willReturn(Optional.empty());
-        given(seniorProfileRepository.findByMemberId(any())).willReturn(Optional.of(seniorProfile));
-        given(seniorProfileRepository.findAllByFamilyId(any())).willReturn(List.of(seniorProfile));
-        given(familyMembershipRepository.findAllByFamilyIdWithGuardian(any())).willReturn(List.of(membership));
-        given(albumRepository.findTopScoredAlbumIdsByMemberIds(anyList(), any()))
-                .willReturn(new SliceImpl<>(List.of()));
+        given(albumRecommendationService.recommendAlbums(any(), any())).willReturn(List.of());
 
         // when
         SeniorHomeCardsResponse response = seniorHomeService.getHomeCards();
@@ -104,9 +81,6 @@ class SeniorHomeServiceTest {
     void 오늘_복용한_스케줄은_taken이_true다() {
         // given
         Member senior = Member.createMember(MemberType.SENIOR, "부모님", "01011112222");
-        Family family = Family.createFamily("FAMA01");
-        SeniorProfile seniorProfile = SeniorProfile.createSeniorProfile(
-                senior, family, "서울시", null, "1110001", null);
 
         MedicineSchedule schedule = mock(MedicineSchedule.class);
         given(schedule.getId()).willReturn(1L);
@@ -124,11 +98,7 @@ class SeniorHomeServiceTest {
         given(heartRateResultRepository.findByMemberId(any())).willReturn(Optional.empty());
         given(healthScheduleRepository.findByMemberIdAndWeek(any(), any(), any())).willReturn(List.of());
         given(walkRepository.findByMemberAndWalkDate(any(), any())).willReturn(Optional.empty());
-        given(seniorProfileRepository.findByMemberId(any())).willReturn(Optional.of(seniorProfile));
-        given(seniorProfileRepository.findAllByFamilyId(any())).willReturn(List.of(seniorProfile));
-        given(familyMembershipRepository.findAllByFamilyIdWithGuardian(any())).willReturn(List.of());
-        given(albumRepository.findTopScoredAlbumIdsByMemberIds(anyList(), any()))
-                .willReturn(new SliceImpl<>(List.of()));
+        given(albumRecommendationService.recommendAlbums(any(), any())).willReturn(List.of());
 
         // when
         SeniorHomeCardsResponse response = seniorHomeService.getHomeCards();
@@ -143,20 +113,13 @@ class SeniorHomeServiceTest {
     void 오늘_건강일정이_없으면_healthSchedule은_null을_반환한다() {
         // given
         Member senior = Member.createMember(MemberType.SENIOR, "부모님", "01011112222");
-        Family family = Family.createFamily("FAMA01");
-        SeniorProfile seniorProfile = SeniorProfile.createSeniorProfile(
-                senior, family, "서울시", null, "1110001", null);
 
         given(memberUtil.getCurrentMember()).willReturn(senior);
         given(medicineScheduleRepository.findByMemberAndStatusWithDetails(any(), any())).willReturn(List.of());
         given(heartRateResultRepository.findByMemberId(any())).willReturn(Optional.empty());
         given(healthScheduleRepository.findByMemberIdAndWeek(any(), any(), any())).willReturn(List.of());
         given(walkRepository.findByMemberAndWalkDate(any(), any())).willReturn(Optional.empty());
-        given(seniorProfileRepository.findByMemberId(any())).willReturn(Optional.of(seniorProfile));
-        given(seniorProfileRepository.findAllByFamilyId(any())).willReturn(List.of(seniorProfile));
-        given(familyMembershipRepository.findAllByFamilyIdWithGuardian(any())).willReturn(List.of());
-        given(albumRepository.findTopScoredAlbumIdsByMemberIds(anyList(), any()))
-                .willReturn(new SliceImpl<>(List.of()));
+        given(albumRecommendationService.recommendAlbums(any(), any())).willReturn(List.of());
 
         // when
         SeniorHomeCardsResponse response = seniorHomeService.getHomeCards();
@@ -170,20 +133,13 @@ class SeniorHomeServiceTest {
     void 걷기_데이터와_기본목표가_없으면_walk는_null을_반환한다() {
         // given
         Member senior = Member.createMember(MemberType.SENIOR, "부모님", "01011112222");
-        Family family = Family.createFamily("FAMA01");
-        SeniorProfile seniorProfile = SeniorProfile.createSeniorProfile(
-                senior, family, "서울시", null, "1110001", null);
 
         given(memberUtil.getCurrentMember()).willReturn(senior);
         given(medicineScheduleRepository.findByMemberAndStatusWithDetails(any(), any())).willReturn(List.of());
         given(heartRateResultRepository.findByMemberId(any())).willReturn(Optional.empty());
         given(healthScheduleRepository.findByMemberIdAndWeek(any(), any(), any())).willReturn(List.of());
         given(walkRepository.findByMemberAndWalkDate(any(), any())).willReturn(Optional.empty());
-        given(seniorProfileRepository.findByMemberId(any())).willReturn(Optional.of(seniorProfile));
-        given(seniorProfileRepository.findAllByFamilyId(any())).willReturn(List.of(seniorProfile));
-        given(familyMembershipRepository.findAllByFamilyIdWithGuardian(any())).willReturn(List.of());
-        given(albumRepository.findTopScoredAlbumIdsByMemberIds(anyList(), any()))
-                .willReturn(new SliceImpl<>(List.of()));
+        given(albumRecommendationService.recommendAlbums(any(), any())).willReturn(List.of());
 
         // when
         SeniorHomeCardsResponse response = seniorHomeService.getHomeCards();
@@ -197,20 +153,13 @@ class SeniorHomeServiceTest {
     void 가족_앨범이_없으면_albums는_빈배열을_반환한다() {
         // given
         Member senior = Member.createMember(MemberType.SENIOR, "부모님", "01011112222");
-        Family family = Family.createFamily("FAMA01");
-        SeniorProfile seniorProfile = SeniorProfile.createSeniorProfile(
-                senior, family, "서울시", null, "1110001", null);
 
         given(memberUtil.getCurrentMember()).willReturn(senior);
         given(medicineScheduleRepository.findByMemberAndStatusWithDetails(any(), any())).willReturn(List.of());
         given(heartRateResultRepository.findByMemberId(any())).willReturn(Optional.empty());
         given(healthScheduleRepository.findByMemberIdAndWeek(any(), any(), any())).willReturn(List.of());
         given(walkRepository.findByMemberAndWalkDate(any(), any())).willReturn(Optional.empty());
-        given(seniorProfileRepository.findByMemberId(any())).willReturn(Optional.of(seniorProfile));
-        given(seniorProfileRepository.findAllByFamilyId(any())).willReturn(List.of(seniorProfile));
-        given(familyMembershipRepository.findAllByFamilyIdWithGuardian(any())).willReturn(List.of());
-        given(albumRepository.findTopScoredAlbumIdsByMemberIds(anyList(), any()))
-                .willReturn(new SliceImpl<>(List.of()));
+        given(albumRecommendationService.recommendAlbums(any(), any())).willReturn(List.of());
 
         // when
         SeniorHomeCardsResponse response = seniorHomeService.getHomeCards();
